// src/lib/proto/RCRSProto.ts
export interface PropertyProto {
  urn: number;
  defined: boolean;
  intValue?: number | undefined;
  boolValue?: boolean | undefined;
  doubleValue?: number | undefined;
  byteList?: Uint8Array | undefined;
  intList?: IntListProto | undefined;
  intMatrix?: IntMatrixProto | undefined;
  edgeList?: EdgeListProto | undefined;
  point2D?: Point2DProto | undefined;
}
// ...
export interface ChangeSetProto {
  changes: ChangeSetProto_EntityChangeProto[];
  deletes: number[];
}

export interface ChangeSetProto_EntityChangeProto {
  entityID: number;
  urn: number;
  properties: PropertyProto[];
}
// ...
export const PropertyProto: MessageFns<PropertyProto> = {
  encode(message: PropertyProto, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.urn !== 0) {
      writer.uint32(8).int32(message.urn);
    }
    if (message.defined !== false) {
      writer.uint32(16).bool(message.defined);
    }
    if (message.intValue !== undefined) {
      writer.uint32(24).int32(message.intValue);
    }
    if (message.boolValue !== undefined) {
      writer.uint32(32).bool(message.boolValue);
    }
    if (message.doubleValue !== undefined) {
      writer.uint32(41).double(message.doubleValue);
    }
    if (message.byteList !== undefined) {
      writer.uint32(50).bytes(message.byteList);
    }
    if (message.intList !== undefined) {
      IntListProto.encode(message.intList, writer.uint32(58).fork()).join();
    }
    if (message.intMatrix !== undefined) {
      IntMatrixProto.encode(message.intMatrix, writer.uint32(66).fork()).join();
    }
    if (message.edgeList !== undefined) {
      EdgeListProto.encode(message.edgeList, writer.uint32(74).fork()).join();
    }
    if (message.point2D !== undefined) {
      Point2DProto.encode(message.point2D, writer.uint32(82).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): PropertyProto {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePropertyProto();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.urn = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.defined = reader.bool();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.intValue = reader.int32();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.boolValue = reader.bool();
          continue;
        }
        case 5: {
          if (tag !== 41) {
            break;
          }

          message.doubleValue = reader.double();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.byteList = reader.bytes();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.intList = IntListProto.decode(reader, reader.uint32());
          continue;
        }
        case 8: {
          if (tag !== 66) {
            break;
          }

          message.intMatrix = IntMatrixProto.decode(reader, reader.uint32());
          continue;
        }
        case 9: {
          if (tag !== 74) {
            break;
          }

          message.edgeList = EdgeListProto.decode(reader, reader.uint32());
          continue;
        }
        case 10: {
          if (tag !== 82) {
            break;
          }

          message.point2D = Point2DProto.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<PropertyProto>, I>>(base?: I): PropertyProto {
    return PropertyProto.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<PropertyProto>, I>>(object: I): PropertyProto {
    const message = createBasePropertyProto();
    message.urn = object.urn ?? 0;
    message.defined = object.defined ?? false;
    message.intValue = object.intValue ?? undefined;
    message.boolValue = object.boolValue ?? undefined;
    message.doubleValue = object.doubleValue ?? undefined;
    message.byteList = object.byteList ?? undefined;
    message.intList = (object.intList !== undefined && object.intList !== null)
      ? IntListProto.fromPartial(object.intList)
      : undefined;
    message.intMatrix = (object.intMatrix !== undefined && object.intMatrix !== null)
      ? IntMatrixProto.fromPartial(object.intMatrix)
      : undefined;
    message.edgeList = (object.edgeList !== undefined && object.edgeList !== null)
      ? EdgeListProto.fromPartial(object.edgeList)
      : undefined;
    message.point2D = (object.point2D !== undefined && object.point2D !== null)
      ? Point2DProto.fromPartial(object.point2D)
      : undefined;
    return message;
  },
};
// ...
export const ChangeSetProto: MessageFns<ChangeSetProto> = {
  encode(message: ChangeSetProto, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.changes) {
      ChangeSetProto_EntityChangeProto.encode(v!, writer.uint32(10).fork()).join();
    }
    writer.uint32(18).fork();
    for (const v of message.deletes) {
      writer.int32(v);
    }
    writer.join();
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ChangeSetProto {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseChangeSetProto();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.changes.push(ChangeSetProto_EntityChangeProto.decode(reader, reader.uint32()));
          continue;
        }
        case 2: {
          if (tag === 16) {
            message.deletes.push(reader.int32());

            continue;
          }

          if (tag === 18) {
            const end2 = reader.uint32() + reader.pos;
            while (reader.pos < end2) {
              message.deletes.push(reader.int32());
            }

            continue;
          }

          break;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<ChangeSetProto>, I>>(base?: I): ChangeSetProto {
    return ChangeSetProto.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ChangeSetProto>, I>>(object: I): ChangeSetProto {
    const message = createBaseChangeSetProto();
    message.changes = object.changes?.map((e) => ChangeSetProto_EntityChangeProto.fromPartial(e)) || [];
    message.deletes = object.deletes?.map((e) => e) || [];
    return message;
  },
};

function createBaseChangeSetProto_EntityChangeProto(): ChangeSetProto_EntityChangeProto {
  return { entityID: 0, urn: 0, properties: [] };
}

export const ChangeSetProto_EntityChangeProto: MessageFns<ChangeSetProto_EntityChangeProto> = {
  encode(message: ChangeSetProto_EntityChangeProto, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.entityID !== 0) {
      writer.uint32(8).int32(message.entityID);
    }
    if (message.urn !== 0) {
      writer.uint32(16).int32(message.urn);
    }
    for (const v of message.properties) {
      PropertyProto.encode(v!, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ChangeSetProto_EntityChangeProto {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseChangeSetProto_EntityChangeProto();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.entityID = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.urn = reader.int32();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.properties.push(PropertyProto.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<ChangeSetProto_EntityChangeProto>, I>>(
    base?: I,
  ): ChangeSetProto_EntityChangeProto {
    return ChangeSetProto_EntityChangeProto.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ChangeSetProto_EntityChangeProto>, I>>(
    object: I,
  ): ChangeSetProto_EntityChangeProto {
    const message = createBaseChangeSetProto_EntityChangeProto();
    message.entityID = object.entityID ?? 0;
    message.urn = object.urn ?? 0;
    message.properties = object.properties?.map((e) => PropertyProto.fromPartial(e)) || [];
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}